+++ src/rhydb/common/rhydb_directory.cpp
@@ -1,4 +1,4 @@
-#include "rhydb/common/silo_directory.h"
+#include "rhydb/common/rhydb_directory.h"
 
 #include <spdlog/spdlog.h>
