+++ src/rhydb/common/rhydb_directory.h
