+++ src/rhydb/common/rhydb_directory.test.cpp
@@ -1,4 +1,4 @@
-#include "rhydb/common/silo_directory.h"
+#include "rhydb/common/rhydb_directory.h"
 
 #include <gtest/gtest.h>
